seperation of api call actions into seperate file
