+++ src/components/Module.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from "react-router-dom";
 
 import TaskList from './TaskList';
