+++ src/containers/Login.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
 import Login from "../components/Login";
-import { authenticate } from "../data/actions";
+import { authenticate } from "../data/actions_API";
 import { onFormElementChange } from "../data/actions";
 
 const mapStateToProps = state => ({
